refactor: extract tabby_common::api::code
mark CodeSearch being Send + Sync
