+++ Cargo.toml
@@ -34,3 +34,4 @@ futures = "0.3.28"
 async-stream = "0.3.5"
 regex = "1.10.0"
 thiserror = "1.0.49"
+utoipa = "3.3"
