+++ crates/tabby-common/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod api;
 pub mod config;
 pub mod events;
 pub mod index;
