+++ crates/tabby/Cargo.toml
@@ -14,7 +14,7 @@ tabby-inference = { path = "../tabby-inference" }
 axum = "0.6"
 hyper = { version = "0.14", features = ["full"] }
 tokio = { workspace = true, features = ["full"] }
-utoipa = { version = "3.3", features = ["axum_extras", "preserve_order"] }
+utoipa = { workspace= true, features = ["axum_extras", "preserve_order"] }
 utoipa-swagger-ui = { version = "3.1", features = ["axum"] }
 serde = { workspace = true }
 serdeconv = { workspace = true }
@@ -42,9 +42,9 @@ axum-streams = { version = "0.9.1", features = ["json"] }
 minijinja = { version = "1.0.8", features = ["loader"] }
 textdistance = "1.0.2"
 regex.workspace = true
-thiserror.workspace = true
 llama-cpp-bindings = { path = "../llama-cpp-bindings" }
 futures.workspace = true
+async-trait.workspace = true
 
 [dependencies.uuid]
 version = "1.3.3"
